% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_etc.R
\name{calc_etc}
\alias{calc_etc}
\title{Calcul de l'evapotranspiration culturale (ETc)}
\usage{
calc_etc(eto, kc)
}
\arguments{
\item{eto}{Vecteur ou valeur de l'ETo journaliere (mm/jour)}

\item{kc}{Vecteur ou valeur du coefficient cultural Kc}
}
\value{
Vecteur de valeurs ETc (mm/jour)
}
\description{
Calcule l'ETc journaliere selon la relation ETc = Kc x ETo
}
\examples{
eto <- c(4.5, 5.0, 4.8, 5.2, 4.9)
kc  <- c(0.4, 0.5, 0.6, 0.7, 0.8)
etc <- calc_etc(eto, kc)
print(etc)

}
