% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eto.R
\name{calc_eto}
\alias{calc_eto}
\title{Calcul de l'Evapotranspiration de Reference (ETo) - Methode FAO-56}
\usage{
calc_eto(tmax, tmin, solar_rad, wind_speed, humidity, altitude = 0)
}
\arguments{
\item{tmax}{Temperature maximale journaliere (C)}

\item{tmin}{Temperature minimale journaliere (C)}

\item{solar_rad}{Rayonnement solaire journalier (MJ/m2/jour)}

\item{wind_speed}{Vitesse du vent a 2m (m/s)}

\item{humidity}{Humidite relative moyenne (\%)}

\item{altitude}{Altitude du site (m). Par defaut 0.}
}
\value{
Valeur numerique de l'ETo en mm/jour
}
\description{
Calcule l'evapotranspiration de reference journaliere selon la methode
de Penman-Monteith simplifiee (FAO-56).
}
\examples{
eto <- calc_eto(tmax = 32, tmin = 18, solar_rad = 22,
                wind_speed = 2.5, humidity = 45)
print(eto)

}
