% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kc.R
\name{calc_kc}
\alias{calc_kc}
\title{Calcul des coefficients culturaux Kc}
\usage{
calc_kc(culture = "ble", date_debut, duree_stades = NULL)
}
\arguments{
\item{culture}{Nom de la culture : "ble", "mais", "tomate",
"olivier", "pomme_de_terre"}

\item{date_debut}{Date de debut de la culture (format "AAAA-MM-JJ")}

\item{duree_stades}{Vecteur de 4 durees en jours pour les stades :
initial, developpement, mi-saison, fin saison}
}
\value{
Un dataframe avec les colonnes : date, stade, Kc
}
\description{
Genere une serie journaliere de coefficients Kc selon les stades
phenologiques de la culture, basee sur les valeurs FAO-56.
}
\examples{
kc <- calc_kc(culture = "ble", date_debut = "2023-11-01",
              duree_stades = c(30, 40, 40, 20))
head(kc)

}
