% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generer un rapport automatique HTML}
\usage{
generate_report(
  bilan,
  planning,
  culture = "ble",
  station = "Station",
  output_file = "rapport_hydroAgri.html"
)
}
\arguments{
\item{bilan}{Dataframe issu de soil_water_balance()}

\item{planning}{Liste issue de irrigation_schedule()}

\item{culture}{Nom de la culture. Par defaut "ble"}

\item{station}{Nom de la station. Par defaut "Station"}

\item{output_file}{Nom du fichier de sortie. Par defaut "rapport_hydroAgri.html"}
}
\value{
Chemin vers le fichier rapport genere
}
\description{
Genere un rapport complet du bilan hydrique en format HTML
contenant les donnees meteo, ETo, ETc, bilan hydrique,
irrigation, graphiques et recommandations.
}
\examples{
\dontrun{
etc   <- runif(90, 2, 5)
rain  <- ifelse(runif(90) > 0.8, runif(90, 5, 20), 0)
bilan <- soil_water_balance(etc, rain, RU = 100)
planning <- irrigation_schedule(etc, rain, RU = 100)
generate_report(bilan, planning, culture = "ble")
}

}
