% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_climate.R
\name{interpolate_climate}
\alias{interpolate_climate}
\title{Interpolation spatiale des donnees climatiques}
\usage{
interpolate_climate(stations_data, variable, methode = "idw", resolution = 0.1)
}
\arguments{
\item{stations_data}{Dataframe avec les colonnes : lon, lat, et la variable}

\item{variable}{Nom de la variable a interpoler (ex: "ETo", "Tmax", "Rain")}

\item{methode}{Methode : "idw" ou "krigeage". Par defaut "idw"}

\item{resolution}{Resolution de la grille en degres. Par defaut 0.1}
}
\value{
Un objet SpatialGridDataFrame (raster interpole)
}
\description{
Interpole spatialement les donnees climatiques ponctuelles vers
une grille reguliere par la methode IDW ou krigeage.
}
\examples{
stations <- data.frame(
  lon = c(-8.0, -7.5, -7.0, -8.5, -7.8),
  lat = c(31.5, 31.8, 32.0, 31.2, 32.2),
  ETo = c(5.2, 4.8, 5.5, 4.5, 5.0)
)
\dontrun{
grille <- interpolate_climate(stations, variable = "ETo")
}

}
