% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irrigation_schedule.R
\name{irrigation_schedule}
\alias{irrigation_schedule}
\title{Calcul du calendrier d'irrigation}
\usage{
irrigation_schedule(
  etc,
  rain,
  RU = 100,
  RFU_ratio = 0.5,
  RU_init = NULL,
  dates = NULL
)
}
\arguments{
\item{etc}{Vecteur des ETc journalieres (mm/jour)}

\item{rain}{Vecteur des precipitations journalieres (mm)}

\item{RU}{Reserve utile totale du sol (mm). Par defaut 100}

\item{RFU_ratio}{Fraction de RU pour la RFU (0 a 1). Par defaut 0.5}

\item{RU_init}{Humidite initiale (mm). Par defaut egale a RU}

\item{dates}{Vecteur de dates. Par defaut NULL}
}
\value{
Une liste avec :
\item{calendrier}{Dataframe du calendrier journalier}
\item{evenements}{Dataframe des irrigations uniquement}
\item{total_irrigation}{Volume total irrigue (mm)}
}
\description{
Determine les dates et doses d'irrigation optimales selon le seuil
de la Reserve Facilement Utilisable (RFU).
}
\examples{
etc  <- runif(90, 2, 6)
rain <- ifelse(runif(90) > 0.85, runif(90, 5, 20), 0)
planning <- irrigation_schedule(etc, rain, RU = 100)
planning$evenements

}
