% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raster_eto.R
\name{load_raster_eto}
\alias{load_raster_eto}
\title{Charger des donnees raster ETo}
\usage{
load_raster_eto(
  file,
  extent_zone = NULL,
  crs_target = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{file}{Chemin vers le fichier raster (.tif, .nc, .grd)}

\item{extent_zone}{Vecteur de 4 valeurs (xmin, xmax, ymin, ymax).
Si NULL, conserve l'etendue originale}

\item{crs_target}{Systeme de coordonnees cible. Par defaut WGS84}
}
\value{
Un objet raster
}
\description{
Importe un fichier raster d'evapotranspiration de reference
(MODIS, WorldClim ou raster local), le reprojette et le decoupe.
}
\examples{
\dontrun{
r <- load_raster_eto("eto_maroc.tif",
                     extent_zone = c(-6, -4, 31, 33))
raster::plot(r)
}

}
