% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_irrigation_map.R
\name{plot_irrigation_map}
\alias{plot_irrigation_map}
\title{Cartographie spatiale des besoins en irrigation}
\usage{
plot_irrigation_map(stations_data, variable = "Irrigation", titre = NULL)
}
\arguments{
\item{stations_data}{Dataframe avec les colonnes :
lon, lat, et la variable a cartographier}

\item{variable}{Variable a cartographier :
"Irrigation", "Deficit", "ETo". Par defaut "Irrigation"}

\item{titre}{Titre de la carte. Par defaut NULL}
}
\value{
Un objet ggplot
}
\description{
Cree des cartes des besoins en irrigation, deficit hydrique
et ETo spatial a partir de donnees de stations.
}
\examples{
stations <- data.frame(
  lon       = c(-8.0, -7.5, -7.0, -8.5, -7.8),
  lat       = c(31.5, 31.8, 32.0, 31.2, 32.2),
  Irrigation = c(120, 95, 140, 88, 110)
)
plot_irrigation_map(stations, variable = "Irrigation")

}
