% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_water_balance.R
\name{plot_water_balance}
\alias{plot_water_balance}
\title{Visualisation du bilan hydrique}
\usage{
plot_water_balance(
  bilan,
  dates = NULL,
  titre = "Bilan Hydrique",
  save_plot = FALSE,
  fichier = "bilan.png"
)
}
\arguments{
\item{bilan}{Dataframe issu de soil_water_balance() avec
les colonnes : ETc, Rain, Irrigation, Stockage, Deficit}

\item{dates}{Vecteur de dates. Par defaut NULL}

\item{titre}{Titre du graphique. Par defaut "Bilan Hydrique"}

\item{save_plot}{Sauvegarder le graphique ? Par defaut FALSE}

\item{fichier}{Nom du fichier de sortie. Par defaut "bilan.png"}
}
\value{
Un objet ggplot
}
\description{
Cree des graphiques du bilan hydrique : pluie, ETo, ETc,
irrigation et humidite du sol.
}
\examples{
etc  <- runif(90, 2, 5)
rain <- ifelse(runif(90) > 0.8, runif(90, 5, 20), 0)
bilan <- soil_water_balance(etc, rain, RU = 100)
plot_water_balance(bilan)

}
