% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_dashboard.R
\name{shiny_dashboard}
\alias{shiny_dashboard}
\title{Dashboard Shiny interactif}
\usage{
shiny_dashboard()
}
\value{
Lance l'application Shiny
}
\description{
Lance une application Shiny interactive pour visualiser
le bilan hydrique et les besoins en irrigation.
}
\examples{
\dontrun{
shiny_dashboard()
}

}
