% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_water_balance.R
\name{soil_water_balance}
\alias{soil_water_balance}
\title{Calcul du bilan hydrique du sol}
\usage{
soil_water_balance(etc, rain, irrigation = 0, RU = 100, RU_init = NULL)
}
\arguments{
\item{etc}{Vecteur des ETc journalieres (mm/jour)}

\item{rain}{Vecteur des precipitations journalieres (mm)}

\item{irrigation}{Vecteur des apports d'irrigation (mm). Par defaut 0}

\item{RU}{Reserve utile totale du sol (mm). Par defaut 100}

\item{RU_init}{Humidite initiale du sol (mm). Par defaut egale a RU}
}
\value{
Un dataframe avec les colonnes :
ETc, Rain, Irrigation, Stockage, Drainage, Deficit
}
\description{
Calcule le bilan hydrique journalier du sol selon la methode FAO-56.
}
\examples{
etc  <- runif(30, 2, 5)
rain <- ifelse(runif(30) > 0.8, runif(30, 5, 20), 0)
bilan <- soil_water_balance(etc, rain, RU = 100)
head(bilan)

}
