% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_water_balance.R
\name{summarize_water_balance}
\alias{summarize_water_balance}
\title{Resume du bilan hydrique}
\usage{
summarize_water_balance(bilan)
}
\arguments{
\item{bilan}{Dataframe issu de soil_water_balance() ou
irrigation_schedule()$calendrier}
}
\value{
Un dataframe avec les statistiques saisonnieres
}
\description{
Calcule les statistiques saisonnieres du bilan hydrique :
ETc totale, irrigation totale, deficit cumule et efficacite.
}
\examples{
etc  <- runif(90, 2, 5)
rain <- ifelse(runif(90) > 0.8, runif(90, 5, 20), 0)
bilan <- soil_water_balance(etc, rain, RU = 100)
resume <- summarize_water_balance(bilan)
print(resume)

}
